[r38477] Remove XvMC support
The functionality was removed from FFmpeg for years.

With the recent FFmpeg 7 release it was completely removed from
FFmpeg, so --enable-xvmc in MPlayer now results in build errors.

This removal was discussed with Ivan who initially developed
XvMC support and maintained it in the last two decades.

(From upstream SVN r38477)